refactor: seperate logic for network
This seperates the logic for getting network details into another
file
